Task 29 crud imoveis

+++ app/Http/Controllers/Admin/TipoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Tipo;
+use App\Imovel;
 
 class TipoController extends Controller
 {
@@ -49,6 +50,27 @@ class TipoController extends Controller
 
   public function deletar($id)
   {
+    /*
+    Teste para verificar se existe imovel
+    Relacionado a essa cidade que quer deletar
+    */
+    if(Imovel::where('tipo_id','=',$id)->count()){
+
+      $msg = "Não é possível deletar esse tipo de imóvel!
+      Esses imóveis(";
+      $imoveis = Imovel::where('tipo_id','=',$id)->get();
+
+      foreach ($imoveis as $imovel) {
+        $msg .= "id:".$imovel->id." ";
+      }
+      $msg.= ") estão relacionados.";
+
+      \Session::flash('mensagem',[
+        'msg'=> $msg,
+        'class' => 'red white-text'
+      ]);
+      return redirect()->route('admin.tipos');
+    }
     Tipo::find($id)->delete();
     \Session::flash('mensagem',[
       'msg'=> 'Registro deletado com sucesso',
